Redraw Kreska ukośna stroke for Ł.
This is per discussion on #24. I have attempted to counter two optical illusions; firstly the right hand half is lowered slightly, and secondly the angle of the underside is adjusted so as not to appear convex. Thus the control points do not form an exact rectangle.

Smallcaps and lowercase ł to follow shortly.