Dredge out character classes.
This uses the same `smallcaps` class for both `c2sc` and `smcp` substitutions; other than for legibility of the feature file, the idea here is to ensure that both the `lowercase` and `uppercase` classes are consistent with respect to each other.

Some substitutions are present in one class, but not the other, as some glyphs do not have an uppercase or lowercase counterpart.

I've (neccessarily) added a few missing substitutions whilst doing this, but there are probably still several missing.

This commit leaves the feature files functional per #29.